# api/controllers/system.js
'use strict';

module.exports.getDevices = (req, res) => {
    global.myq.getDevices()
        .then( (devices) => {
            res.json( { data: devices, result : 'ok'  } );
        })
        .catch( (err) => {
            res.status(500).json( { code: err.code || 0, message: err.message } );
        });
};

module.exports.Reload = (req, res) => {
    global.myq.Reload()
        .then( () => {
            res.json( { data: {}, result : 'ok'  } );
        })
        .catch( (err) => {
            res.status(500).json( { code: err.code || 0, message: err.message } );
        });
};

module.exports.getDeviceStatus = (req, res) => {
    global.myq.getDeviceStatus(req.swagger.params.id.value)
        .then( (status) => {
            res.json( { data: { status: status }, result : 'ok' } );
        })
        .catch( (err) => {
            res.status(500).json( { code: err.code || 0, message: err.message } );
        });
};